refactoring service agents
