+++ src/Mantasflowers.Services/ServiceAgents/Firebase/IFirebaseServiceAgent.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using Mantasflowers.Contracts.ServiceAgents.Firebase.Request;
 using Mantasflowers.Contracts.ServiceAgents.Firebase.Response;
 
-namespace Mantasflowers.Services.ServiceAgents
+namespace Mantasflowers.Services.ServiceAgents.Firebase
 {
     public interface IFirebaseServiceAgent
     {
