+++ src/Mantasflowers.Services/ServiceAgents/ParcelMonkey/IParcelMonkeyServiceAgent.cs
@@ -1,0 +1,6 @@
+﻿namespace Mantasflowers.Services.ServiceAgents.ParcelMonkey
+{
+    public interface IParcelMonkeyServiceAgent
+    {
+    }
+}
