+++ src/Mantasflowers.Services/ServiceAgents/ParcelMonkey/ParcelMonkeyServiceAgent.cs
@@ -1,0 +1,6 @@
+﻿namespace Mantasflowers.Services.ServiceAgents.ParcelMonkey
+{
+    public class ParcelMonkeyServiceAgent : IParcelMonkeyServiceAgent
+    {
+    }
+}
